feat(@spartacus/storefront): Remove hardcoded links and context selectors from the header (#890)
While we not yet have backend support for links and context selectors (language, currency), we have finalised the header UI part of those elements. We've done this with a temporary interceptor that adds the intended CMS content to the OCC response. 
The context selector is now CMS component and can be extended for other contexts.