feat(@spartacus/storefront): Map cx-login-form to cms component (#1457)
The `ReturningCustomerLoginComponent` is mapped to the `cx-login-form` component. We also moved the register link logic for new users into this component, as we intend to drop hardcoded layout logic. With these changes, we've dropped the hardcoded LoginPageModule and rely on generic ContentPage logic only. As a side effect, we removed the NotAuthGuard in this case, and rely on the component logic to redirect the user in case of already logged in.

closes #1455, #1456