feat(@spartacus/storefront): add image resolver for og:image meta tag (#1605)
In order to share pages with social media, such as facebook, twitter, pinterest, etc. it is important to provide the right image in the meta tags. The Open Graph standard from Facebook is widely adopted for this purpose. 
In this PR we've added the og:image tag for the PDP.

closes #1598