feat(core): config mapping from key to namespace (translation chunk name)
Previously, keys defined in the html templates contained the name of the file with lazy loaded chunk (we call it namespace). It coupled tightly the html templates with the structure of lazy loaded chunks of translations. This could lead into problems when a customer wants to rearrange chunks (i.e. merge translations related to two features into one chunk).

So we now use keys without namespaces in the templates and we map them to namespaces by config. If no mapping provided, the first part of the key is treated as the namespace.

fixes GH-1894