refactor(core, storefront): remove mechanism for transofrmation of links from CMS (#1994)
The feature of recognition and transformation of links from CMS was built before we had generic CMS content pages. Now we would like to be more CMS-driven and depend on the links from CMS, when we can. So the feature of transformation links from CMS is now redundant and being removed.

fixes GH-1976

BREAKING CHANGE:
Recognizing and transforming dynamic urls from CMS is not supported anymore - `{ url: ... } | cxTranslateUrl` will not work.