docs: document data binding logic (#1996)
In order to adapt other systems for backends, we've introduced connector logic. The connector logic is used to adapt OCC as well as other systems. This is related to the bigger story of data binding, from UI to backend. The documentation provided with this ticket focus on data binding in general, and the ability to adapt other systems in Spartacus.

closes #1748