Re-render components when they change (#2223)
While Single Page Applications can leverage the advantage of caching the page structure after first load and use it during the session or even in sub sequential visits, content isn't necessarily static and could change based on the user interaction. The storefront should reload content while the user re-visits a page.

In the future we could optimise the (re)loading of CMS content, but this should be mainly a backend related story: An optimised interaction from front-end to backend could be implemented to avoid duplicated calls for content. One idea that comes to mind is to provide a hash (in backend) of the page structure and pass this in sub sequential calls. The backend could evaluate the hash and return a 304 (not modified) in case of equal page structure. This would save the network bandwidth and client-site processing of the data.

Closes GH-1538
Closes GH-1920