refactor(core, storefront): rename 'cxPath' to 'cxRoute' and drop support of 'cxRedirectTo' (#2337)
The `cxRedirectTo` data property of `Route` was created to redirect from old links to standard Spartacus ones. We've get rid of those redirects in #1978, so we remove the support of this property.

The `data.cxPath` is now ranamed to `data.cxRoute`

closes  GH-2090

BREAKING CHANGES:
Configurable route's `data.cxPath` property is now renamed to `data.cxRoute`. Dropped support of the `cxRedirectTo` property.