fix: redirectUrl logic and move it from ngrx to service (#2578)
After login/register we need to redirect to the last visited page. Unfortunatelly, in some cases it didn't work. Now it's fixed. 

Moreover, the logic of redirectUrl is now moved from router reducer to separate service. Now the urls preserving the redirect url are not only `/login` and `/login/register` (as was hardcoded before), but every page with the `NotAuthGuard`.

By the way, the error after login (GH-2693) was fixed.

close GH-2541, fix GH-2693