Dynamic page layout handling (#2866)
Page Layout service will be able to filter page layout by optional PageLayoutHandler. The PageLayoutHandler will be able to modify the template given by the backend (or static config).

One use case for this is the dynamic building of the cart page; while the cms provides statically a paragraph in case there's no cart entries, we cannot always show this. The PageLayoutHandler would be able to change this based on the cart entries.

Closes GH-2661