fix: move RouterModule.forRoot up to StorefrontModule (#3193)
The import of `RouterModule.forRoot()` it's moved up to the `StorefrontModule` which can be eaisly re-build by a developer. This allows for easy providing custom options for the Angular's `RouterModule`.

The `SiteContextModule` also had to be moved to the `StorefrontModule`, after the `RouterModule`. When it was imported deeper in the modules tree, its `SiteContextUrlSerializer` was shadowed by the RouterModule's `DefaultUrlSerializer`.

Removed also obsolete imports of our `RoutingModule` and providing `RoutingService` (which is `providedIn: 'root'`).

close GH-3088