Initial implementation for Structured Data (#4882)
Adds a design for adding json-ld data into Spartacus. The implementation includes product and breadcrumb schema's, but can be further extended. The schema's aren't added by default, since the builders are provided in `JsonLdBuilderModule`, which is not imported by default in any of the storefront modules. We might do this in an upcoming minor release. 
Additionally, a low-level level `json-ld` directive was added, which can be used to add schema fragments for specific components. 

Closes #4748