feat (styles): introduce theme driven page templates (#5248)
Similar to the theming we use for components (i.e. calydon), we can now use a theme for a page layouts. The page layout can have addition placeholder selectors, so that whenever the theme is used, additional styles rules are added to the final CSS.

Additional, we introduced a white and blacklist of page templates so that customers can more easily opt-out of certain page template styles. The variables are $page-template-blacklist and $page-template-whitelist.

Moreover, the page templates can now be imported separately, using @import '@spartacus/styles/scss/page-template';

closes #5159