Performance improvements for Product load  (#5046)
The product detail payload is currently huge and used in multiple places. For example, in the product carousel, we only use the name, price and primary image, but the FULL data set is loaded. Ideally, we break up the product load into pieces, so we can gradually load data.

Closes #3666

Optimize occ product adapter to take advantage of changes introduced in #3666

Closes #5024