fix: Category Aware Search Profile not applied on Category Page (#5510)
When using Adaptive Search only Global profiles were applying, but not
category profiles.  This is because Adaptive Search expects the use of
the allCategories field (either as a Filter Query or a Facet Value) to
mark a category page query.

This affects both categories and brands (previouly using the category
and brand facets respectively).  Both have been changed to query on
the allCategories field.

The JSP storefronts all use the alLCategories field for all category
page requests & do not differentiate between category/brand/etc when
querying.

closes #5510