fix: Category Aware Search Profile not applied on Category Page (#5510)
Fix CategoryPageMetaResolver to understand the allCategories facet.

Note: Commerce setup needs modifying for this to work correctly.  The
allCategories facet by default has no Facet Value Display Name Provider
set.  Without that the breadcrumb disaplys the code rather than the name.
Set this to the same as the category facet (categoryFacetDisplayNameProvider)
to have proper names in the breadcrumb