fix: Category Aware Search Profile not applied on Category Page (#6096)
When using Adaptive Search only Global profiles were applying, but not category profiles. This is because Adaptive Search expects the use of the `allCategories` field (either as a Filter Query or a Facet Value) to mark a category page query.

This affects both categories and brands (previously using the category and brand facets respectively). Both have been changed to query on the `allCategories` field.

Note: Commerce setup needs modifying for this to work correctly. The `allCategories` facet by default has no Facet Value Display Name Provider set. Without that the breadcrumb displays the code rather than the name. Set this to the same as the category facet (categoryFacetDisplayNameProvider) to have proper names in the breadcrumb

BREAKING CHANGE:

Spartacus requires the `allCategories` index property begin configured on the backend. The `categories` index property is no longer used.

fixes #5510

Co-authored-by: tobi-or-not-tobi <tobi-or-not-tobi@users.noreply.github.com>