feat: configurable pagination (#6060)
The pagination component is lower level component which is reused in various lists, including the Product Listing Page. With this PR we're solving various issues and increase the configurability and extensibility of the component. The default pagination in for Spartacus is also changed from 2.0 onwards to only show the page range (with maximum 3 pages) and a begin and end navigation link: `« 4 (5) 6 »`. 

BREAKING CHANGE:

The pagination component is refactored completely:
- The UI supports anchor links next to action links
- The pagination links are build by a builder class
- The links, link labels and link positions are fully configurable
- The pagination is fully configurable by CSS as well
- The pagination is hidden when there is no pagination needed

This closes #6060, fixes #3045, fixes #5631, fixes #4632, fixes #5487, fixes #3856, fixes #3118