fix: Revert breaking changes in consignment tracking (#6525)
Closes #6524 

BREAKING CHANGES:
With 1905.4 version endpoint for consignment tracking changed due to a bug. We introduced in spartacus change in the default endpoint configuration and new parameter (userId) in adapter/connector method `getConsignmentTracking` to support this version and above. If you extended that functionality make sure to always pass it (we also fallback to `current` user when `userId` is not provided). If you work with older versions of the platform to fix the consignment tracking feature you have to change endpoint for consignemtTracking in configuration to `orders/${orderCode}/consignments/${consignmentCode}/tracking