feat: Support business driven facet top values (#6597)
The search business tools allow for so-called topValues configuration, a number that indicates the direct visible values of a facet. So far, this number was only hard-coded in Spartacus. With this feature, the business user can customise the number per facet.

The topValues are normalised from the backend to the UI model. A default number (6) is used in case there are no top values given.

closes #5692