feat: introduce breakpoint methods to indicate breakpoint appearance (#6961)
Adding 3 methods to the breakpoint service:

isUp – Indicates whether the given breakpoint is equal or larger than the window size
isDown – Indicates whether the current screen size is larger than the minimum size of the given breakpoint
isEqual – Indicates whether the current screen size fits to the given breakpoint
Additionally, we've fixed a glitch in the breakpoint service, which is causing a breaking change.

closes #6960

BREAKING CHANGE:

The `BreakpointService` contained an issue in version 1.0. An incorrect breakpoint as used, which had an effect on the layout configuration. The `b2cLayoutConfiguration` has been adapted to use the correct breakpoints:
- the layout configuration for the header section is using large (`lg`) instead of the incorrect `md`. 
- the layout configuration for the navigation section is using large (`lg`) instead of the incorrect `md`. 

If you're using the `BreakpointService` directly in your project, you should be aware of this change.