fix: provide Session affinity with the withCredentials=true flag on http requests (#7139)
This could be considered a feature as well. It's been introduced in 2.0, but a backport was required for customers who scale out their API. In order to configure this behaviour (which is turned of by default), you must provide the interceptor in a module (i.e. in the AppModule)

`providers: [withCredentialsInterceptorProvider]`

Additionally, a CORS configuration is required in the backend to pass requests from another domain. The following configuration must be installed:

`corsfilter.ycommercewebservices.allowCredentials=true`

This can be done either in the backoffice or by using project.properties.

closes #7104