Enable Storefront to utilize checkout groups
This is just an example how the storefront lib could be modified in order to enable Spartacus  to use different flavours of checkouts. The concept of checkout groups is in SAP Commerce has a long history associated with accelerator and it would be nice to have that feature available in the Spartacus.

From backed perspective there is already information holding default checkout group defined on BaseStore level so it's easy to extend occ /basesites endpoint to include in the response information pulled from base store. I can also provide example of that modification if needed

Proposed changes are just entry point to further discussion related to techniques for extending checkout flow and making it dependent on either site configuration or dynamic information collected from basket

From the system integrator perspecive it's important to stay as close to original libraries as possible to avoid maintenance overhead so the code needs to be open for extension. In the provided scaffolding CheckoutFlowConfigService implementation storefront could rely on information that comes from the ng store to adjust checkout group to desired one by the customer. The checkout group could be then modified in dynamic fashion by updating the site state using custom reducers.

Example use case:
- customer requested to add gift wrap service on basket so additional step (alternative checkout flavour) could be presented where customer can enter gift recipient
- customer buy areosol with alcohol and he needs to validate age

The next step after different group might be attempt to model checkout in dynamic fashion by setting step condition in which business logic for deciding if the step is in or out of scope in the checkout flow (but this is separate use case)

Sap configurable checkout docs:
https://help.sap.com/viewer/4c33bf189ab9409e84e589295c36d96e/1905/en-US/8acb4f6f86691014be81d4246aa1c9c5.html

Sap dynamic checkout docs in telco accelerator:
https://help.sap.com/viewer/4c33bf189ab9409e84e589295c36d96e/1905/en-US/a6612fad4cfe40c694b3b7e44376e520.html?q=Dynamic%20Checkout%20Flows