fix: avoid emitting empty value from TranslationService when used in APP_INITIALZIER (#7207)
When the translation service was used too early (i.e. in APP_INITIALIZER), it emitted an empty value, because the i18next engine was not yet initialized. Now the service can be used safely in the APP_INITIALIZER, because the translated values are emitted only after the i18next is ready.

closes #3229
closes #3838