feat: MediaService takes media threshold for a given media format (#7230)
The media service no longer depends on the breakpoints, and is now using configurable media size information. It will now resolve a media always, either by configuration, best match, or the first format, to avoid an empty image for banners or products.

It is no longer needed to pass in a specific format to the MediaComponent, but it won't break if you still do so. Spartacus will not pass in a format anymore.

closes #7229, fixes #6629, fixes #5398, fixes #5395