feat: Traverse parent nodes rather than parent elements in outlet directive (#7342)
The outlet directive uses the host element to render the original template or any additional UI. The host element is read by traversing the DOM until there's an HtmlElement is found. The API to use is parentElement, but this is blocking projects who need to support ie11 (which Spartacus doesn't). To make life easier, we make an exception in the code base and find the parentNode instead.

closes #7220