fix: fix: don't dispatch context change action on initial set
During the first visit in spartacus storefront the CMS page data was loaded twice for no good reason. Now it's loaded once.

The bug was caused by the pageRefresh$ effect which was triggered on each langauge change (even on the first one - when the langauge was initially set).

Now we don't dispatch the langauge change action on setting the initial value. So CMS data is not loaded twice on initial page visit.

Before, the langauge (and currency) sometimes was intialized multiple times by the competing sources of truth (i.e. routing, session storage, default config), which caused a fake language change event in some circumstances (combinations of configuraiton and user journey).

Now it's orchiestrated, so the langauge is initialized from once, using one source (the first matched condition wins):

If SSR transferred state is present, take value from there. OR
If param was configured to be persisted from URL, take value from there. (When the value is not present in the current URL, update the URL using default configured value and then use this value.) OR
If value can be retrieved from session storage, take value from there OR
Otherwise take the static default value from config
The same priorities hold for currency. But base site have: 1, 2 and 4 (no session storage as source).

fixes GH-6900 , fixes GH-7306, fixes GH-7307