feat: Version specific styles to avoid breaking changes (#7956)
We're cautious with introducing changing styles in a minor release, as it would conflict the semantic versioning model. In order to improve the style layer during minor releases a $minorVersion is introduced. With this approach, the former style theme is being sunset (and will be repurposed going forward).

You must explicitly opt-in to a certain (minor) version, to benefit from the latest style rules. Alternatively, yo can use the $useLatestStyles flag to always use the latest version. This is convenient for demos and POCs.

closes #7680