feat: split view feature (#8077)
The split-view component allows to navigate through different views while the initial views are still accessible. You can navigate through the views, by opening and closing views. The split view comes with a CSS implementation that is using slide transitions when opening/closing views.

An additional deactivate guard is provided in those cases where views are route based. The guard adds a small delay to the routing, so that the transition can finish in time.

close #8076