sync epic (#8126)
* feat: Table component (#8044) (#8047)

Generic table component, that allows for breakpoint configurable table columns. The table is rendered with template based headers and data elements, so that container components can influence the rendering of the table. Customers can use the same technique to apply additional rendering.

close #8044

* Move content of tech changes page to docs repo (#8057)

Closes GH-8056

* fix: occ-endpoints.service should work properly with config initializers

Closes #8067

* GH-7526 Marking save for later e2e test as flaky for now

* feat: split view feature (#8077)

The split-view component allows to navigate through different views while the initial views are still accessible. You can navigate through the views, by opening and closing views. The split view comes with a CSS implementation that is using slide transitions when opening/closing views.

An additional deactivate guard is provided in those cases where views are route based. The guard adds a small delay to the routing, so that the transition can finish in time.

close #8076

* small refactor of SplitViewComponent (#8085)

Small improvements

* chore: add a secondary entry point to the product lib (#8080)

This PR adds a secondary entry point for the `@spartacus/product` library.

Importing from the the _Product_ library can now be done in the following way:

- `import {...} from '@spartacus/product/configuration` - exposes API relevant only to the _product configuration_
- `import {...} from '@spartacus/product'` - exposes all APIs - currently only from `@spartacus/product/configuration` but in the future it might contain more.

* fix: remove brands from active filters (#8083)

Closes #7806 

Co-authored-by: tobi-or-not-tobi <tobi-or-not-tobi@users.noreply.github.com>

Co-authored-by: tobi-or-not-tobi <tobi-or-not-tobi@users.noreply.github.com>
Co-authored-by: Matthew <43279967+gladius-mtl@users.noreply.github.com>
Co-authored-by: Stan <dunqan@w9.pl>
Co-authored-by: Gilberto Alvarado <gilberto.alvarado.jimenez@sap.com>
Co-authored-by: Nikola Zarić <nikola.zaric.ns@gmail.com>
Co-authored-by: Kacper Knapik <kknapik@divante.pl>