Fix memory leak and allow init to be customized.
Constructors can't be overwritten and should never had code directly in it. Otherwise implementation teams can't change any of the logic. 
In this case, the constructor also has a memory leak due to subscription without unsubscribe. This will not be an issue for a single user given the active-cart service won't be initialize more than once, but in SSR, it will result in memory leak between sessions.