feat: support bi-direction icons (rtl vs ltr) (#8394)
Some icons should flip based on the direction. Icons are typically ltr oriented, however the implementation is not opinionated. Any icon can be configured to be flipped at rtl or ltr. A `flip-at-ltr` or `flip-at-rtl` class is added to the icon host element, indicating for which direction it should be flipped. 

The list of flippable icons can be configured in the icon configuration, using the `DirectionMode.RTL` or `DirectionMode.LTR`.

```typescript
icon: {
    flipDirection: {
      STAR: DirectionMode.RTL,
      CARET_RIGHT: DirectionMode.RTL,
      CARET_LEFT: DirectionMode.RTL,
    },
  },
```

closes #8392