feat: Add page template to the root storefront element (#8450)
Introduces the page template as a css class on the `cx-storefront` element. This is done by introducing a `APP_BOOTSTRAP_LISTENER` that adds the page template class to the root component.

The former css styles add layout related style rules to the page template class, which is why we make the CSS styles more specific to be applied only to the cx-page-layout. To avoid breaking changes, this feature is introduced with a feature level flag (2.1). If you enable 2.1 feature flag, the new CSS styles will change from `.LandingPage2Template` to `cx-page-layout.LandingPage2Template`.

closes #8450