chore: avoid duplicated config in b2b.feature and b2c.feature (#8953)
Currently some config is duplicated, which makes it harder to maintain. Common part should be extracted to one common place. Moreover, default translations should be configured in one place, but not re-declared in all feature-recpipes (in case of using multiple of them at once), which would make the config chunks unnecesarily bigger and slower to load (deep merge).

closes #8952