feat: provide observable context$ inside OutletContextData (#9431)
The `cxOutlet` directive can provide the template context via the `cxOutletContext` input. This context is currently passed to the outlet child components via Injection Token `OutletContextData`. But the context in this token is static and doesn't allow for reacting to changes of the original `cxOutletContext` input.

Now we expose also new observable property of `OutletContextData`: `context$`, which emits on each context change.

closes GH-9233