fix - add ellipsis in sub-lists of MyCompany when content too long (#9518)
The CSS ellipsis was working well for the main tables (i.e. the main users table), but was not working for the sub lists (i.e. list of Users for the Unit). When the sub-list's cell had too long content, text went beyond the borders.

Now it's fixed. The last column is sqhashed and aligned to the right. And the rest of preceding columns take proportionally he remaining space and are truncated with ellipsis, if their content is too long. 

Additional improvements:
- added `title` attributes to the cell components - for better desktop experience, to allow lookup of the full value
- reused templates in cell-components to avoid duplication
- dropped obsolete padding styling for single-cell link (it caused the unit address list to have inconsistent padding)
- added missing padding to the unit cell (before that, the ellipsis were squeezed too close to the unit cell)
- fix: aligned lines of ghost design (before they were at the very bottom of the sub-list cell)
- fix: marked 'unit' column as non-linkable - in unit's assigned approvers sub-list (before they were clickable and pointed to empty URL)

closes GH-9490