fix: avoid sharing i18next instance in between SSR requests (for 3.0.x) (#10746)
The i18next global instance was shared in between all SSR requests. This was buggy, i.e. when 2 requests are made at the same time for different acitive langauges. Now we fix it, by creating a fresh instance for each SSR request (for each app bootstrap).

backport to 3.0.x

close #8100