fix: Make use of userIdService to determine user Id for OCC call (GH-10787)
Instead of taking the user from the current session cart, it is determined from userIdService. This allows to run product configurators also in the ASM scenario. 
For the textfield configurator, the addToCart and configureFromCart process runs fine now.

Note however that for the rulebased configurator, a backend access check still prevents the proper cart integration in the ASM scenario.

Closes GH-10787