feat: product configurator cpq support (#11001)
Adds support for the CPQ configurator to feature library product-configurator. The CPQ configurator re-uses most of the existing components, facades and store artefacts. It adds some new components for handling of bundles, and it comes with a connection layer that calls CPQ directly for the interactive configuration process, instead of connecting via OCC like the other configurators do.