fix: Fix OnNavigateFocusService with SSR v2 (#13305)
In backport #13285 to 3.4.x we added a SSR check to avoid the error caused by the absence of the method `scrollIntoView()` in SSR:
```
ERROR TypeError: (...).scrollIntoView is not a function
```

We didn't want to add a new constructor dependency to avoid a breaking change. So instead of injecting proper PLATFORM_ID or WindowRef, we checked `if (typeof window === 'undefined')`. 

However we realized this solution might not work for customers that mock `window` object in SSR. (Note: we don't recommend modking `window` in SSR, but for some customers it might be needed.)

So this PR removes the window check and gracefully handles the absence of the method `scrollIntoView()`, by adding `?.` check.

The agreed tradeoff is that the subscriptions will run in SSR, although they don't need to. 

closes #13221