feat: use error details of the form control to populate translation params in `FormErrorsComponent` (#13382)
Previously, i18n params were  passed to the component `cx-form-errors` only via the component input `translationParams`, which forces parent components (that own the form) to pass the `translationParams` down the components tree. This was impacting negatively the API of the components in the middle which unnecessarily had to implement the `translationParams` input and pass it down the tree.

This is now avoided as we start using the error details object in the `cx-form-errors`. This error is embedded in the form control (which is already passed down the components tree).

Note: for backward compatibility the `translationParams` input is preserved and takes precedence over the params from the error details.

closes #13375