fix: add option for reusing a pending SSR render for subsequent similar requests (instead of fallback to CSR) (#13680)
Before this PR:
When the html for a request was being rendered and in the meantime new requests arrived for the same rendering key (e.g. for the same URL), they were resolved immediately with a CSR fallback.

This PR preserves the previous default behavior, but also adds a new opt-in config that helps avoiding unnecessary CSR fallbacks.

The new opt-in config `reuseCurrentRendering` allows the subsequent requests to wait for the result of the pending render that was triggered by the 1st request. Finally, the all the waiting requests for the same rendering key (e.g. URL) will reuse the result of a single completed render.

Note 1: Same as previously, each request can time out independently (according to the `timeout` config) and then this request is resolved with a CSR fallback. However, with the new config `reuseCurrentRendering` enabled, other requests (which didn't time out yet) can still wait for the result of the render.

Note 2: The existing config `concurency` limits the number of the pending renders. So when config `reuseCurrentRendering` is enabled,  requests for the same rendering key take up just one concurrency slot, because they share just one pending render.

Note 3: Previously the `RenderinStrategy.ALWAYS_SSR` caused triggering of a new render for each request, even if there was a pending render for the same rendering key. This PR preserves the previous default behavior. However, with the new config `reuseCurrentRendering` enabled, the subsequent requests for the same key won't trigger new renders, but will wait for the pending render to complete.

closes #13623