feat: Provide reduced header layout for configurator pages (#13949)
Provides a reduced layout, which does not contain elements like search and miniCart, in order to not distract the end-user while configuring. 
This change is optional and needs to explicitly activated in a customer's app module as of 4.x. The internal app activates it by importing the respective layout modules in the VC and CPQ feature modules.
closes #13807