feat: standard page header if configurator in display mode (#14339)
The interactive and overview configurator page show a reduced header, in order to not distract the user from the configuration task. 
In case however the configurator page is navigated from the order history or checkout screen (in display-only mode), it should behave like its context and display the standard SPA header.   

Closes #14154
Closes #14153