feat: introduce floating add to cart button (#14452)
The styling of the add to cart button has been changed. The container where the button is in now requires the entire length of the window. Its position has been set to 'fixed'. When users scroll up, the button follows. It will stick to the bottom section of the screen, just under price summary once the price summary is in the view port. 

Closes #14390