test: update test visa to 12/2027 expiry (#16036)
The cypress test began failing as the expiry date was set to 07/2022.
When we reached 2022-07-01, the tests began to fail. Setting the expiry to
the far future fixes this.

A date later than 2027 was not chosen as that would require scrolling a
dropdown during checkout, which is unreliable in headless mode in a pipeline.

(cherry picked from commit 8e76316a4057b6c2502323dab2750a3342986144)