feat: bind anonymous cart to a registered user in the ASM interface (#15776)
The new feature to bind an anonymous cart to an emulated user is available after starting a
customer emulation session.  A new component will appear in the ASM header that provides an input
and "Bind" button.  The input will by default display the current bound cart (if there is one) and allow
the assisted service agent to enter a new cart ID to bind using the "Bind" button.

New code includes a new facade/adapter/connector set for the cart bind operation, a new
loader component (dot-spinner) currently intended for use in ASM UI, and a new component for 
the ASM header to bind a cart.

A known issue is that the new styles for 5.1 are temporarily activated by default until a future fix to
style libraries that will make the base style version level work with feature libraries.