chore: Use "development" configuration for "yarn dev:ssr"  (#16417)
Now for `yarn dev:ssr` we use the "development" configuration in `angular.json`, which generates sourcemaps and disables optimizations (e.g. minification) of code. Thanks to that now it's much easier to debug SSR, while running `yarn dev:ssr`. Now when you `Attach to a Node process` (to Node process of `yarn dev:ssr`) e.g. in VSCode and set a breakpoint for a lines in the source code, the control is paused at this line, when SSR request is processed.

closes CXSPA-1672