fix: unhandled account summary attachment download when timed out (#16575)
When the session is timed out and the user tries to download an attachment from the account summary page, the UI was not redirecting the user to the login page.

This was happening because the auth interceptor was looking for a specific error type in the json body of the error message. In the case of attachment api, the endpoints returns data type blob. For this reason, the error message was unreadable by the auth interceptor.

The fix is to add a blob interceptor that runs before the auth interceptor. This interceptor checks if the type is blob, and if so extracts the contents and rethrows the error as json. This error is then caught by the auth interceptor and handled correctly.

Ticket: https://jira.tools.sap/browse/CXSPA-1766

Co-authored-by: Weizheng Gao <weizheng.gao@sap.com>
Co-authored-by: Giancarlo Cordero Ortiz <46171897+giancorderoortiz@users.noreply.github.com>
Co-authored-by: Giancarlo Cordero Ortiz <giancarlo.cordero.ortiz@sap.com>