Fix: Avoid immediate navigation to conflict group (#16634)
When Commerce 22.05 or newer is used the UI immediately navigated to the conflict solver as soon as a conflict appears. This is due to additional data in the response from commerce. This bug is fixed with this correction. The  search for the first group with attributes has been adapted to the new response data.