fix: ensure ids used for name attributes and keys for cxFocus directive are unquie (#16727)
In case of immediate conflict resolution an attribute may get rendered twice, once in the conflict resolution popup and once still in background. As the id generation assumes that an attribute is only rendered once, this leads to duplicate ids. In most cases this does not cause any issues, however there are 2 situations where ids need to be unique.
*  key of cxFocus directive
*  name attributes of controls
for these cases there is an additional uiContextKey injected into the generated ids.