fix: 'ERROR TypeError: changes.config is undefined' in focus directive (#16785)
The ngOnChanges hook of the focus directive falsely assumes that the parameter changes always has a property config. In case the property is missing, ERROR TypeError: changes.config is undefined is logged into the browser console. Other branches in the directive, for example see line 109, correctly use an optional chaining operator to access this property. This PR adds the optional chaining to line 72, as well, so that no error is logged anymore.
For reproducing the issue, please see the description of https://jira.tools.sap/browse/CXSPA-2193

fixes: https://jira.tools.sap/browse/CXSPA-2193